Honor excludedFiles in the config file
The current code tramples the excludedFiles setting in the config file
with either the value supplied on the command line or an empty array.
This fixes that; additionally, if excludedFiles is specified in both
the config file and on the command line, the union of both is used.

Tests have been updated/added to cover the different scenarios.