Add a max-warnings cli flag
Where number below `max-warnings` result in a zero exit code. This is
desirable for projects that allow a certain amount of warnings to pass
the build system, and a build system that relies on a zero exit code
for the build to pass.

We default to 0, for backwards compatibility. A max warning of -1 will
allow any number of warnings to pass.

Closes #149