fix: check +5 numbers for spaceAroundOperator (#433)
Update lesshint to work with shellscape/postcss-values-parser#45, which
will return a number for +5 instead of an operator and a number.

So, check both operators and numbers, ignore any numbers that don’t
start with an operator (+-*/), allow the node to be the last node (no
nextElement), and use a whitespace regex to check the none spacing
instead of the exact value.